# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/menu/restaurant/topics": {
            "get": {
                "description": "Get list of topics by restaurant ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topic"
                ],
                "summary": "Get topics by restaurant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Topic name keyword",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get topics by restaurant successfully",
                        "schema": {
                            "$ref": "#/definitions/app.GetTopicsByRestaurantSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            }
        },
        "/api/menu/topics": {
            "post": {
                "description": "Create a new topic with name, price, description, and optional image or logo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topic"
                ],
                "summary": "Create a new topic",
                "parameters": [
                    {
                        "description": "Topic data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/topicapp.CreateTopicRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create topic success",
                        "schema": {
                            "$ref": "#/definitions/app.CreateTopicSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            }
        },
        "/api/menu/topics/combobox": {
            "get": {
                "description": "Get list of topics for combobox by restaurant (optional parent_id)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topic"
                ],
                "summary": "Get topic combobox",
                "responses": {
                    "200": {
                        "description": "Get topic combobox successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/app.GetTopicsByRestaurantComboboxSuccessResponseDoc"
                            }
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            }
        },
        "/api/menu/topics/{id}": {
            "get": {
                "description": "Get detailed information of a topic using its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topic"
                ],
                "summary": "Get topic by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Topic ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get topic successfully",
                        "schema": {
                            "$ref": "#/definitions/app.GetTopicSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            },
            "put": {
                "description": "Update topic information by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topic"
                ],
                "summary": "Update topic",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Topic ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Topic data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/topicapp.UpdateTopicRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update topic successfully",
                        "schema": {
                            "$ref": "#/definitions/app.UpdateTopicSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete topic by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topic"
                ],
                "summary": "Delete topic",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Topic ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete topic successfully",
                        "schema": {
                            "$ref": "#/definitions/app.DeleteTopicSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.CreateTopicSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
        "app.DeleteTopicSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
        "app.GetTopicSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
        "app.GetTopicsByRestaurantComboboxSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Combobox"
                    }
                },
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
        "app.GetTopicsByRestaurantResponseDoc": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/topicapp.GetTopicResponse"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total_items": {
                    "type": "integer",
                    "example": 42
                },
                "total_pages": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "app.GetTopicsByRestaurantSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/app.GetTopicsByRestaurantResponseDoc"
                },
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
        "app.UpdateTopicSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
        "response.Combobox": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                },
                "value": {}
            }
        },
        "response.ErrorDoc": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
        "topicapp.CreateTopicRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
        "topicapp.GetTopicResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "restaurant_id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
        "topicapp.UpdateTopicRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
...
    }
}